feat: Add endpoint to fetch confirmed enrollments by class
This commit adds a new endpoint, `pegaMatriculasPorTurma`, to the `PessoaController` class. This endpoint allows fetching all confirmed enrollments for a specific class. The logic is implemented using the `findAndCountAll` method provided by the `database.Matriculas` model. If an error occurs during the process, it is handled by returning an appropriate error response.

The corresponding route, `/pessoas/matricula/:turmaId/confirmadas`, is added to the `pessoasRoute.js` file. This route maps to the `pegaMatriculasPorTurma` method in the `PessoaController`.

These changes enhance the functionality of the API by providing the ability to retrieve confirmed enrollments for a particular class.