Add functionality to retrieve information of full classrooms
This commit adds a new method `pegaTurmasLotadas` to the `PessoaController` that retrieves the information of full classrooms. The method uses `findAndCountAll` to query the database for confirmed registrations per classroom and returns the count of classrooms with a count greater than or equal to a predefined `lotacaoTurma`. The method is accessible via a new GET route `/pessoas/matricula/lotada` in the `pessoasRoute` file.