Add cancel feature for individuals and their associated registrations
This commit adds a new method in the `PessoaController` class called `cancelaPessoa`. This method is responsible for canceling an individual's registration and marking them as inactive. It updates the `ativo` field in the `Pessoas` table to `false` and sets the `status` field in the `Matriculas` table to "cancelado" for the corresponding individual. The commit also includes the necessary route configuration in the `pessoasRoute.js` file.

The commit message is as follows:

"Add cancel feature for individuals and their associated registrations"