feat: Add Services class for fetching data dynamically
This commit introduces a new Services class that allows fetching data dynamically based on the provided model name. The class is added in the Services.js file and includes a constructor and a method, pegaTodosOsRegistros, which retrieves all records from the specified model. This functionality is utilized in the PessoaController.js file, where instead of directly calling the database.Pessoas.findAll() method, the new services instance is used to fetch all active personnes.